succes realise second page

+++ app/src/main/java/com/example/testapplication/network/remotedatasource/ApiRepository.kt
@@ -5,5 +5,5 @@ import com.example.testapplication.network.data_station.StationResponse
 
 interface ApiRepository {
     suspend fun login(login:String,password:String): AuthByLoggingResponse
-    suspend fun getStations(latitude:Float,longitude:Float,limit:Int):List<StationResponse>
+    suspend fun getStations(latitude:Double,longitude:Double,limit:Int):List<StationResponse>
 }
